[IMP] web_tour: Allow customization of rainbow man message
Before this commit, the rainbow man message at the end of tours is fixed
to a hardcoded message.

This commit allows to customize the message within the definition of the
tour, to add the ability to use tailored messages depending on the app
the tour is about. If no custom message is set, the old default will
still be used.

Task ID: 2241549

closes odoo/odoo#49779

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>