[FIX] web: pivot view nocontenthelp message
Since the commit https://github.com/odoo/odoo/commit/7e3e6b92e3e27c3727a0430e05d5364b015814eb
when help is not defined in action then nocontenthelp is displayed as
[object, object] due to the desciption passed from renderder of the
pivot view.

So in this commit while calling the NoContentHelper pass the
descitption as string instead of jquery element.
we can use the t-value in template but that string will create
issue for transalation so pass the desctiption from the pivot render.

closes odoo/odoo#50347

Taskid: 2240167
Closes: #49859
X-original-commit: 6fa058c4ac9901cfe4afcf5ce24e3e1bcb86f6f3
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>