[FIX] point_of_sale,pos_cash_rounding: fix change error popup on checkout
Configure your POS with bank payment method only
Make a sale with rounding necessary (i.e. 1,29€)
The amount will be rounded. Go through the payment screen.

Error message will appear because no payment methods available is
capable of handling the change, but this is not necessary since we are
rounding.

Splitting the code to add a more specific condition in pos_cash_rounding
fix the issue

opw-2239946

closes odoo/odoo#50216

X-original-commit: 6fb2e84561d1593a88b019d10ce916a4168c0cf1
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>