[FIX] account: don't raise an error when using account controls together with section/notes lines on invoices
Section and note invoice lines are now account.move.line objects with a distinct value in display_type field. When restricting the use of accounts on a payable/receivable journal and making an invoice in it with at least one section or note, and only using legal accounts, it always failed, as the note/section account was empty, hence not in the allowed accounts.

closes odoo/odoo#50297

X-original-commit: a2e22b06b446fbc881677091812924ece436e0f5
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>