[FIX] mail: avoid scrolling "NavBar" when dropdown open
Before this commit, when a SysTray dropdown is open the user can scroll
the element behind the dropdown.

After this commit, we disable the scroll until we close the dropdown.

Steps to reproduce:
* Open Odoo on "Mobile"
* Open the Activity SysTray
* Scroll the "NavBar" menu (=> Bug)

Note this fix is also applied in Desktop, as for me the "NavBar"
shouldn't never scroll on Desktop.

Task ID: 2231956
Task ID: 2234042

closes odoo/odoo#51716

X-original-commit: 7bb155991b9daec37edf73ef90cbad26268d76bb
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: rfr-odoo <rfr-odoo@users.noreply.github.com>